code improvement for water cell 2 * changed Integer.MAX_VALUE to Integer.MAX_VALUE * 1000L for creative cells
