+++ src/main/java/com/glodblock/github/common/item/ItemInfinityWaterStorageCell.java
@@ -3,21 +3,29 @@ package com.glodblock.github.common.item;
 import static com.glodblock.github.util.Util.FluidUtil.water_bucket;
 
 import java.util.EnumSet;
+import java.util.List;
 
+import net.minecraft.client.gui.GuiScreen;
+import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.inventory.IInventory;
 import net.minecraft.item.ItemStack;
+import net.minecraft.util.StatCollector;
 
 import com.glodblock.github.FluidCraft;
-import com.glodblock.github.api.FluidCraftAPI;
+import com.glodblock.github.common.storage.IFluidCellInventory;
+import com.glodblock.github.common.storage.IFluidCellInventoryHandler;
 import com.glodblock.github.common.storage.IStorageFluidCell;
 import com.glodblock.github.common.tabs.FluidCraftingTabs;
 import com.glodblock.github.loader.IRegister;
 import com.glodblock.github.util.NameConst;
 
+import appeng.api.AEApi;
 import appeng.api.config.FuzzyMode;
+import appeng.api.storage.IMEInventoryHandler;
+import appeng.api.storage.StorageChannel;
 import appeng.api.storage.data.IAEFluidStack;
 import appeng.core.features.AEFeature;
-import appeng.items.contents.CellUpgrades;
+import appeng.tile.inventory.AppEngInternalInventory;
 import cpw.mods.fml.common.registry.GameRegistry;
 
 public class ItemInfinityWaterStorageCell extends ItemBaseInfinityStorageCell
@@ -28,12 +36,49 @@ public class ItemInfinityWaterStorageCell extends ItemBaseInfinityStorageCell
         return new InfinityConfig(water_bucket);
     }
 
+    public static class InfinityConfig extends AppEngInternalInventory {
+
+        public InfinityConfig(final ItemStack is) {
+            super(null, 1);
+            this.setInventorySlotContents(0, is);
+        }
+
+        @Override
+        public void markDirty() {}
+    }
+
     public ItemInfinityWaterStorageCell() {
         super();
         setUnlocalizedName(NameConst.ITEM_INFINITY_WATER_FLUID_STORAGE);
         setTextureName(FluidCraft.resource(NameConst.ITEM_INFINITY_FLUID_STORAGE).toString());
         this.setFeature(EnumSet.of(AEFeature.StorageCells));
         this.setMaxStackSize(1);
+    }
+
+    @Override
+    public void addCheckedInformation(final ItemStack stack, final EntityPlayer player, final List<String> lines,
+                                      final boolean displayMoreInfo) {
+        final IMEInventoryHandler<?> inventory = AEApi.instance().registries().cell()
+            .getCellInventory(stack, null, StorageChannel.FLUIDS);
+
+        if (inventory instanceof final IFluidCellInventoryHandler handler) {
+            final IFluidCellInventory cellInventory = handler.getCellInv();
+
+            if (GuiScreen.isCtrlKeyDown()) {
+                if (!cellInventory.getContents().isEmpty()) {
+                    lines.add(StatCollector.translateToLocal(NameConst.TT_CELL_CONTENTS));
+                    for (IAEFluidStack fluid : cellInventory.getContents()) {
+                        if (fluid != null) {
+                            lines.add(String.format("  %s %s", StatCollector.translateToLocal(NameConst.TT_INFINITY_FLUID_STORAGE_TIPS),  fluid.getFluidStack().getLocalizedName()));
+                        }
+                    }
+                } else {
+                    lines.add(StatCollector.translateToLocal(NameConst.TT_CELL_EMPTY));
+                }
+            } else {
+                lines.add(StatCollector.translateToLocal(NameConst.TT_CTRL_FOR_MORE));
+            }
+        }
     }
 
     @Override
@@ -49,51 +94,11 @@ public class ItemInfinityWaterStorageCell extends ItemBaseInfinityStorageCell
     }
 
     @Override
-    public long getBytes(ItemStack cellItem) {
-        return 0;
-    }
-
-    @Override
-    public int getBytesPerType(ItemStack cellItem) {
-        return 0;
-    }
-
-    @Override
-    public boolean isBlackListed(ItemStack cellItem, IAEFluidStack requestedAddition) {
-        return requestedAddition == null || requestedAddition.getFluid() == null
-                || FluidCraftAPI.instance().isBlacklistedInStorage(requestedAddition.getFluid().getClass());
-    }
-
-    @Override
-    public boolean storableInStorageCell() {
-        return false;
-    }
-
-    @Override
-    public boolean isStorageCell(ItemStack i) {
-        return true;
-    }
-
-    @Override
-    public double getIdleDrain(ItemStack is) {
-        return 0;
-    }
-
-    @Override
-    public int getTotalTypes(ItemStack cellItem) {
-        return 0;
-    }
-
-    @Override
-    public IInventory getUpgradesInventory(ItemStack is) {
-        return new CellUpgrades(is, 0);
-    }
-
-    @Override
     public FuzzyMode getFuzzyMode(ItemStack is) {
         return FuzzyMode.IGNORE_ALL;
     }
 
     @Override
     public void setFuzzyMode(ItemStack is, FuzzyMode fzMode) {}
+
 }
